WIP add new effects, reads, writes, to abx interface
